added flatpak check before installing

+++ core/tabs/applications-setup/virtualization/gnome-boxes.sh
@@ -16,7 +16,9 @@ installBoxes() {
             fi 
             ;;
         *)
-            "$ESCALATION_TOOL" flatpak install --noninteractive org.gnome.Boxes
+            if command_exists flatpak; then
+                "$ESCALATION_TOOL" flatpak install --noninteractive org.gnome.Boxes
+            fi
             ;;
     esac
 }
